app/bin/requirements.txt
moviepy
opencv-python
Pillow
